Extract refresh sensor
